### 2018-07-31 #23 @whedon
[PRE REVIEW]: An Introduction to Applied Bioinformatics: a free, open, and interactive text.
**Submitting author:** @gregcaporaso (<a href="http://orcid.org/0000-0002-8865-1670">James Gregory Caporaso</a>)
**Repository:** <a href="https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics/blob/master/paper.md" target ="_blank">https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics/blob/master/paper.md</a>
**Version:** 0.1.3-dev
