### 2018-07-31 #23 @labarba
@tracykteal This submission is squarely on your field. Given the extent of the materials, I suggest you find three reviewers and divvy up the content for them to review. All could check installation, and so on, but could focus on just a section of the content.
