### 2018-07-31 #23 @tracykteal
@astrobiomike would you be available to review? We can divide the content into sections for the pedagogy/content, so perhaps you could pick a section to review.
