### 2018-08-02 #23 @AstrobioMike
@tracykteal

Absolutely! Looking forward to it :) I'll post here with thoughts on specific section(s) after I get to look over things more closely
