### 2018-08-08 75ca5da
adds co-author Matthew Dillon
Matt Dillon is helping with modernizing the installation process for this project and various build-related tasks.
