### 2018-08-08 3834fe4
TEST: update install process to use environment.yml
