### 2018-08-09 f5e01ad
testing less restrictive skbio version
