### 2018-08-09 869a1ad
TEST: update install process to use environment.yml (#286)
