### 2018-08-09 003f0c0
TEST: replace deprecated runipy test with jupyter nbconvert
