### 2018-08-09 e01131e
TEST: replace deprecated runipy test with jupyter nbconvert (#287)
