+++ book/exercises/multiple-sequence-alignment.md
@@ -3,7 +3,7 @@
 
 ## Purpose <link src="t1fQKW"/>
 
-In this assignment you'll use multiple sequence alignment to reconstruct the phylogeny of a group of organisms based on their 16S rRNA sequences. This assignment builds on ideas from the previous assignment, in that in the last assignment you were identifying good primers to use for amplifying 16S from diverse organisms, and in this assignment we're using those sequences to group organisms by their relatedness. Because of the very large numbers of sequences that are commonly obtained in a modern DNA-sequencing-based experiment, grouping similar sequences and then working with representative sequences for each of those groups is common for computational efficiency. We'll be exploring these ideas in more detail through-out the next segments of the class.
+In this assignment you'll use multiple sequence alignment to reconstruct the phylogeny of a group of organisms based on their 16S rRNA sequences. This assignment builds on ideas from the previous assignment, in that in the last assignment you were identifying good primers to use for amplifying 16S from diverse organisms, and in this assignment we're using those sequences to group organisms by their relatedness. Because of the very large numbers of sequences that are commonly obtained in a modern DNA-sequencing-based experiment, grouping similar sequences and then working with representative sequences for each of those groups is common for computational efficiency. We'll be exploring these ideas in more detail throughout the next segments of the class.
 
 From a bioinformatics standpoint, we usually start working with sequence in fasta format, very similar to the sequences in the cell below. See [here](http://www.bioinformatics.nl/tools/crab_fasta.html) for an explanation of the fasta format.
 
@@ -25,7 +25,7 @@ Continue to work with IPython Notebooks and interact with python code. Understan
 
 ## Functions that you will need to complete the exercise. <link src="vZtErB"/>
 
-Remember to learn about what a function does you can run:
+Remember to learn about what a function does, you can run:
 
 `help(name_of_function)`
 
@@ -157,7 +157,7 @@ Display the guide tree for the sequences in the cell below.
 
 ## Question 3 <link src="8xf4kF"/>
 
-What are the differences in the guide tree from *Question 2*, the tree that is generated after 1 iterations of iterative multiple sequence alignment, and the tree that is generated after 5 iterations of iterative multiple sequence alignment? Display the trees for both 1 and 5 iterations of iterative multiple sequence alignment.
+What are the differences in the guide tree from *Question 2*, the tree that is generated after 1 iteration of iterative multiple sequence alignment, and the tree that is generated after 5 iterations of iterative multiple sequence alignment? Display the trees for both 1 and 5 iterations of iterative multiple sequence alignment.
 
 ```python
 >>> help(iterative_msa_and_tree)
