### 2018-08-18 #23 @labarba
@tracykteal You were looking for a second reviewer for this submission, correct? Once you identify one, use `@whedon add @username as reviewer` and _then_ start the review (that way you get two checklists in the review issue, automatically).
