### 2018-08-28 #23 @gregcaporaso
Hi @tracykteal and @labarba, has review started for this manuscript? Thanks for the update.
