### 2018-08-28 531f6e3
Update pairwise-alignment.md
added line on first time variables are stored; reminder of `%psource` at first use; note that it's ok if it's confusing
