### 2018-08-28 b490082
Update pairwise-alignment.md
typo
