### 2018-08-28 f85d9d3
Update pairwise-alignment.md
printing altered seqs while going along to help keep the mental picture
