### 2018-08-28 f70a44f
Update pairwise-alignment.md
typo
