### 2018-08-28 e74c3af
Update pairwise-alignment.md
typo; minor edits
