### 2018-08-29 #27 @whedon
[REVIEW]: An Introduction to Applied Bioinformatics: a free, open, and interactive text.
**Submitting author:** @gregcaporaso (<a href="http://orcid.org/0000-0002-8865-1670">James Gregory Caporaso</a>)
**Repository:** <a href="https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics" target ="_blank">https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics</a>
**Version:** 0.1.3-dev

- [x] **Version:** Does the release version given match the repository release (0.1.3-dev)?
