### 2018-08-29 #27 @tracykteal
It seems like something in the paper.md isn't allowing it to compile, but it's not clear what it is. I'll see if there's more information in error logs.
