### 2018-08-30 992e3de
Update pairwise-alignment.md
added note about running initial imports again if you return to the middle of the chapter later; changed `show_table` to `show_F` in text
