### 2018-08-30 #27 @arfon
> It seems like something in the paper.md isn't allowing it to compile, but it's not clear what it is. I'll see if there's more information in error logs.

I think the issue is with the URL for the repository. This should be the root of the Git repository, i.e.:

```
https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics

not

https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics/blob/master/paper.md
```
