+++ book/fundamentals/pairwise-alignment.md
@@ -3,7 +3,7 @@
 
 One of the most fundamental problems in bioinformatics is determining how "similar" a pair of biological sequences are. There are many applications for this, including inferring the function or source organism of an unknown gene sequence, developing hypotheses about the relatedness of organisms, or grouping sequences from closely related organisms. On the surface this seems like a pretty straight-forward problem, not one that would have been at the center of decades of research and the subject of [one of the most cited papers](http://scholar.google.com/citations?view_op=view_citation&hl=en&user=VRccPlQAAAAJ&citation_for_view=VRccPlQAAAAJ:u-x6o8ySG0sC) in modern biology. In this chapter we'll explore why determining sequence similarity is harder than it might initially seem, and learn about *pairwise sequence alignment*, the standard approach for determining sequence similarity.
 
-> Here we are loading the tools we are going to use into our current environment. If you return to a section in this chapter later, be sure to run this block of code again.
+> Here we are importing the functions we are going to use into our current working environment. If you return to a section in this chapter later, be sure to run this block of code again.
 
 ```python
 >>> %pylab inline
@@ -12,6 +12,16 @@ One of the most fundamental problems in bioinformatics is determining how "simil
 >>> import numpy as np
 >>> from IPython.core.display import HTML
 >>> from IPython.core import page
+>>> from scipy.spatial.distance import hamming
+>>> from skbio import DNA, Protein
+>>> from iab.algorithms import show_F, show_T, blosum50, show_substitution_matrix, format_dynamic_programming_matrix, format_traceback_matrix
+>>> from skbio.alignment._pairwise import _compute_score_and_traceback_matrices, _traceback, _init_matrices_sw
+>>> from skbio.sequence import Protein
+>>> from skbio.alignment import TabularMSA, global_pairwise_align, local_pairwise_align, local_pairwise_align_nucleotide, local_pairwise_align_ssw
+>>> import timeit
+>>> import pandas as pd
+>>> import seaborn as sns
+
 >>> page.page = print
 ```
 
@@ -19,12 +29,13 @@ Imagine you have three sequences - call them ``r1``and ``r2`` (*r* is for *refer
 
 
 ```python
->>> from scipy.spatial.distance import hamming
->>> from skbio import DNA
-...
 >>> r1 = DNA("ACCCAGGTTAACGGTGACCAGGTACCAGAAGGGTACCAGGTAGGACACACGGGGATTAA")
 >>> r2 = DNA("ACCGAGGTTAACGGTGACCAGGTACCAGAAGGGTACCAGGTAGGAGACACGGCGATTAA")
 >>> q1 = DNA("TTCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCA")
+...
+>>> print(r1)
+>>> print(r2)
+>>> print(q1)
 ```
 Here we just stored 3 sequences as variables. Now we're going to compute the hamming distance between them using the function `hamming()`. Don't forget we can look at what any function does by using `%psource`. Let's see what the `hamming()` function looks like (don't worry if some parts of it aren't entirely clear just yet):
 
@@ -117,7 +128,7 @@ Let's define two sequences, ``seq1`` and ``seq2``, and develop an approach for a
 
[…]
