### 2018-08-30 6eb0b25
Update pairwise-alignment.md
typo
