### 2018-08-30 cd3f707
Update pairwise-alignment.md
typo
