+++ book/fundamentals/pairwise-alignment.md
@@ -8,7 +8,7 @@ One of the most fundamental problems in bioinformatics is determining how "simil
 ```python
 >>> %pylab inline
 >>> from __future__ import division, print_function
-
+...
 >>> import numpy as np
 >>> from IPython.core.display import HTML
 >>> from IPython.core import page
@@ -21,7 +21,7 @@ One of the most fundamental problems in bioinformatics is determining how "simil
 >>> import timeit
 >>> import pandas as pd
 >>> import seaborn as sns
-
+...
 >>> page.page = print
 ```
 
@@ -56,10 +56,10 @@ Here we've assumed that only *substitution events* have occurred, meaning one DN
 
 ```python
 >>> q2 = DNA("TCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCAT")
-
+...
 print(q1)
 print(q2)
-
+...
 >>> print(hamming(r1, q2))
 ```
 
@@ -69,10 +69,10 @@ What we'd really want to do is have a way to indicate that a deletion seems to h
 
 ```python
 >>> q3 = DNA("-TCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCA")
-
+...
 print(q1)
 print(q3)
-
+...
 >>> print(hamming(r1, q3))
 ```
 
