### 2018-08-30 6483f38
Update pairwise-alignment.md
removing more mid-text imports
