### 2018-08-30 d7eedf2
Update pairwise-alignment.md
adding note on time to run examples of aligning increasingly long sequences, and printing a message when done
