### 2018-08-30 662566e
Update pairwise-alignment.md
typo
