### 2018-08-30 5bf13ae
Update database-searching.md
typo
