### 2018-08-30 8260eb5
Update database-searching.md
typo
