+++ book/fundamentals/database-searching.md
@@ -14,6 +14,24 @@ Whose genome is the above sequence encoded in? What is its function? Take a minu
 
 In the context of database searching, a query sequence and a reference sequence that we hypothesize to be homologous can be identical to one another, or they can differ as a result of mutation events. When sequences differ, we're often then interested in how much they differ, or their pairwise similarity, which can help us identify the most closely related of several homologs in the reference database. There is an important distinction in the terms *homology* and *similarity*: homology is a discrete variable, and similarity is a continuous variable. A pair of biological sequences either *are* or *are not* derived from a common ancestor, but they can be more or less similar to each other. Saying that two sequences are 80% homologous doesn't make sense. What people generally mean when they say this is that two sequences are 80% similar, and as a result they are hypothesizing homology between the sequences.
 
+> Here we are going to import all functions used in this chapter. If you want to return somewhere in the middle later, be sure to run this block of code.
+
+```python
+>>> from skbio.alignment import local_pairwise_align_ssw
+>>> from IPython.core import page
+>>> import qiime_default_reference as qdr
+>>> import skbio
+>>> import random
+>>> from iab.algorithms import local_alignment_search, heuristic_local_alignment_search_random, heuristic_local_alignment_search_kmers, heuristic_local_alignment_search_gc, shuffle_sequence, generate_random_score_distribution, fraction_better_or_equivalent_alignments
+>>> import time
+>>> import pandas as pd
+>>> import itertools
+>>> import seaborn as sns
+>>> import IPython.display
+>>> import collections
+>>> import functools
+```
+
 ## Defining the problem <link src="P7kHdy"/>
 
 As mentioned above, if we want to perform a homology search we'll have one or more *query sequences*, and for each we want to know which sequence(s) in a reference database they are most similar to.
@@ -21,7 +39,6 @@ As mentioned above, if we want to perform a homology search we'll have one or mo
 Sequence homology searching can be implemented in a few ways. In this chapter, we'll use the local alignment function that we worked with in [the Pairwise Alignment chapter](alias://a76822), ``local_pairwise_align_ssw``, run it many times to search one *query* sequence against many *reference* sequences, and investigate the highest scoring alignment(s) to identify the best database match. Remember that you can always get help with a function by passing it as an argument to ``help``:
 
 ```python
->>> from skbio.alignment import local_pairwise_align_ssw
 >>> help(local_pairwise_align_ssw)
 ```
 
@@ -38,14 +55,10 @@ First, let's load Greengenes into a list of ``skbio.DNA`` sequence objects, and 
 ```python
 >>> %pylab inline
 ...
->>> from IPython.core import page
 >>> page.page = print
 ```
 
 ```python
->>> import qiime_default_reference as qdr
->>> import skbio
-...
 >>> # Load the taxonomic data
 ... reference_taxonomy = {}
 >>> for e in open(qdr.get_reference_taxonomy()):
@@ -76,8 +89,6 @@ Next, we'll just inspect a couple of the sequences we loaded. Notice how the spe
 For the sake of runtime, we're going to work through this chapter using a random sample of sequences from this database. Here we'll use Python's [random module](https://docs.python.org/3/library/random.html) to select sequences at random.
 
 ```python
->>> import random
-...
 >>> reference_db = random.sample(reference_db, k=5000)
 >>> print("%s sequences are present in the subsampled database." % len(reference_db))
 ```
@@ -122,8 +133,6 @@ Let's define a homology search function that aligns each provided query sequence
 Spend a minute looking at this function to understand what it's doing.
 
 ```python
->>> from iab.algorithms import local_alignment_search
-...
 >>> %psource local_alignment_search
 ```
 
[…]
