### 2018-08-30 c599413
Update database-searching.md
for someone so careful with the use of "homology" above (rightly so), i found it odd you were using "evolutionary distance" to talk about the distance we're calculating between sequences (after much of the pairwise-distance chapter was showing that these are fallible metrics we are using to infer information, haha). just a thought
