### 2018-08-30 9c75a58
Update database-searching.md
some typos
