### 2018-08-30 2279818
Update database-searching.md
typos and some suggested edits
