### 2018-08-30 #27 @AstrobioMike
**Overall comments**
I've gone through some sections in great detail (listed below), and I've looked through the rest briefly so far. Everything I've seen is very nicely done and already really well polished. And the content is stellar. For those sections I have gone through in detail, I have no problems with anything and could only think of very few suggestions that might improve things or help with clarity (to which the pull requests are linked below). Though I consider nothing I've suggested as necessary for me to happily offer my support for publication. This is clearly a well put together resource that will help many aspiring and current bioinformaticians (I'm certainly enjoying learning from it). I'm happy to continue going through sections in detail and providing my minor edits/suggestions to the repository if they are helpful, but in my opinion publication shouldn't wait until I have "fine-toothed" everything – I've certainly seen nothing I would consider problematic in the sections I've glanced through but just haven't gone through in a detailed fashion yet, and to wait for me to do so kind of feels like just waiting for a proofreader. But again that's just my opinion. Here are the sections I've covered in detail so far and links to the pull requests.

**Sections covered in detail and pull requests to very minor edits/suggestions**
Fundamentals: [Pairwise-sequence alignments](https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics/pull/289)
Fundamentals: [Sequence homology searching](https://github.com/caporaso-lab/An-Introduction-To-Applied-Bioinformatics/pull/290)
