### 2018-08-30 d8680c9
ENH: adds new machine learning chapter
