+++ book/fundamentals/machine-learning.md
@@ -1,0 +1,199 @@
+# Machine learning in bioinformatics <link src="7SFnTr"/>
+
+## Supervised v unsupervised classification
+
+## Training data, test data, and cross validation
+
+## Naive Bayes classifiers
+
+```python
+>>> %pylab inline
+...
+>>> from IPython.core import page
+>>> page.page = print
+...
+>>> import pandas as pd
+>>> import skbio
+>>> import numpy as np
+>>> import itertools
+>>> import collections
+```
+
+```python
+>>> import qiime_default_reference as qdr
+...
+>>> # Load the taxonomic data
+... reference_taxonomy = {}
+>>> for e in open(qdr.get_reference_taxonomy()):
+...     seq_id, seq_tax = e.strip().split('\t')
+...     reference_taxonomy[seq_id] = seq_tax
+...
+>>> # Load the reference sequences, and associate the taxonomic annotation with
+... # each as metadata
+... reference_db = []
+>>> for e in skbio.io.read(qdr.get_reference_sequences(), format='fasta', constructor=skbio.DNA):
+...     seq_tax = reference_taxonomy[e.metadata['id']]
+...     e.metadata['taxonomy'] = seq_tax
+...     reference_db.append(e)
+...
+>>> print("%s sequences were loaded from the reference database." % len(reference_db))
+```
+
+```python
+>>> reference_db[0]
+```
+
+```python
+>>> reference_db[-1]
+```
+
+```python
+>>> reference_db = np.random.choice(reference_db, 500, replace=False)
+>>> print("%s sequences are present in the subsampled database." % len(reference_db))
+```
+
+```python
+>>> alphabet = ['A', 'C', 'G', 'T', 'N']
+>>> word_length = 2
+...
+>>> def compute_W(alphabet, word_length):
+>>>  return set(map(''.join, itertools.product(alphabet, repeat=word_length)))
+...
+>>> W = compute_W(alphabet, word_length)
+>>> len(W)
+```
+
+```python
+>>> pd.Series(reference_db[0].kmer_frequencies(k=word_length), name=reference_db[0].metadata['id'])
+```
+
+```python
+>>> def get_taxon_at_level(taxon, level):
+...     taxon = [l.strip() for l in taxon.split(';')]
+...     return '; '.join(taxon[:level])
+```
+
+```python
+>>> import pandas as pd
+...
+>>> taxonomic_level = 2
+>>> word_length = 7
+>>> alphabet = ['A', 'C', 'G', 'T', 'N']
+...
+>>> W = compute_W(alphabet, word_length)
+...
+>>> per_sequence_word_counts = []
+>>> for reference_sequnence in reference_db:
+...     taxon = get_taxon_at_level(reference_sequnence.metadata['taxonomy'], taxonomic_level)
+...     word_counts = dict.fromkeys(W, 0)
+...     word_counts.update(reference_sequnence.kmer_frequencies(k=word_length))
+...     per_sequence_word_counts.append(pd.Series(word_counts, name=taxon))
+>>> # if W is passed as the index can I drop the fromkeys step above?
+... per_sequence_word_counts = pd.DataFrame(data=per_sequence_word_counts).fillna(0).T
+>>> per_sequence_word_counts
+```
+
+```python
+>>> def compute_training_table(per_sequence_word_counts):
+...     N = len(per_sequence_word_counts) # number of training sequences
+...
+...     # number of sequences containing word wi
+...     n_wi = per_sequence_word_counts.astype(bool).sum(axis=1)
+...     n_wi.name = 'n(w_i)'
+...
+...     # probabilities of observing each word
+...     Pi = (n_wi + 0.5) / (N + 1)
+...     Pi.name = 'P_i'
+...
[…]
