+++ book/fundamentals/index.yaml
@@ -4,3 +4,4 @@ contents:
  - multiple-sequence-alignment
  - phylogeny-reconstruction
  - sequence-mapping-and-clustering
+ - machine-learning
