### 2018-08-31 #27 @tracykteal
Thanks @AstrobioMike! Given that the checklist items apply to the lessons as a whole, and you're seeing general consistency throughout, I don't think we need another reviewer for other sections. I'll just give those sections a read through as a final check.
