+++ book/fundamentals/machine-learning.md
@@ -22,6 +22,8 @@ Let's jump in...
 
 The first classifier we'll explore is the popular and relatively simple Naive Bayes classifier. This classifier uses Bayes Theorem to determine the most likely label for an unknown input based on a probabilistic model is has constructed from training data. (_The preceding text needs work._) The model that is constructed is based on user-defined features of the sequences. The most commonly used features for sequence classification tasks such as this is overlapping [kmers](alias://C7hMX5).
 
+We'll begin by importing some libraries that we'll use in this chapter, and then [preparing our reference database and query sequences as we did previously](alias://gAKBxE).
+
 ```python
 >>> %pylab inline
 ...
@@ -63,24 +65,42 @@ The first classifier we'll explore is the popular and relatively simple Naive Ba
 >>> reference_db[-1]
 ```
 
+We'll select a random subset of the reference database to work with here.
+
 ```python
 >>> reference_db = np.random.choice(reference_db, 500, replace=False)
 >>> print("%s sequences are present in the subsampled database." % len(reference_db))
 ```
 
+The first thing our Naive Bayes classifier will need is the set of all possible words of length ``k``. This will be dependent on the value of ``k`` and the characters in our alphabet (i.e., the characters that we should expect to find in the reference database). This set is referred to as ``W``, and can be computed as follows. Given the following alphabet, how many kmers of length 2 are there (i.e., 2-mers)? How many 7-mers are there? How many 7-mers are there if there are twenty characters in our alphabet (as would be the case if we were working with protein sequences instead of DNA sequences)?
+
 ```python
->>> alphabet = ['A', 'C', 'G', 'T', 'N']
->>> word_length = 2
+>>> alphabet = ['A', 'C', 'G', 'T']
+>>> k = 2
 ...
->>> def compute_W(alphabet, word_length):
->>>  return set(map(''.join, itertools.product(alphabet, repeat=word_length)))
+>>> def compute_W(alphabet, k):
+>>>  return set(map(''.join, itertools.product(alphabet, repeat=k)))
 ...
->>> W = compute_W(alphabet, word_length)
+>>> W = compute_W(alphabet, k)
 >>> len(W)
 ```
 
+scikit-bio provides methods for identifying all kmers in a ``skbio.DNA`` sequence object, and for computing the kmer frequencies. This information can be obtained for one of our reference sequences as follows:
+
 ```python
->>> pd.Series(reference_db[0].kmer_frequencies(k=word_length), name=reference_db[0].metadata['id'])
+>>> kmers = reference_db[0].iter_kmers(k=k)
+>>> for kmer in kmers:
+...     print(kmer, end=' ')
+```
+
+```python
+>>> print(reference_db[0].kmer_frequencies(k=k))
+```
+
+This information can be convenient to store in a pandas ``Series`` object:
+
+```python
+>>> pd.Series(reference_db[0].kmer_frequencies(k=k), name=reference_db[0].metadata['id'])
 ```
 
 ```python
@@ -93,48 +113,48 @@ The first classifier we'll explore is the popular and relatively simple Naive Ba
 >>> import pandas as pd
 ...
 >>> taxonomic_level = 2
->>> word_length = 7
+>>> k = 7
 >>> alphabet = ['A', 'C', 'G', 'T', 'N']
 ...
->>> W = compute_W(alphabet, word_length)
+>>> W = compute_W(alphabet, k)
 ...
->>> per_sequence_word_counts = []
+>>> per_sequence_kmer_counts = []
 >>> for reference_sequnence in reference_db:
 ...     taxon = get_taxon_at_level(reference_sequnence.metadata['taxonomy'], taxonomic_level)
[…]
