+++ book/fundamentals/pairwise-alignment.md
@@ -23,8 +23,17 @@ Imagine you have three sequences - call them ``r1``and ``r2`` (*r* is for *refer
 >>> r2 = DNA("ACCGAGGTTAACGGTGACCAGGTACCAGAAGGGTACCAGGTAGGAGACACGGCGATTAA")
 >>> q1 = DNA("TTCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCA")
 ...
+>>> print(r1)
+>>> print(r2)
+>>> print(q1)
+```
+Here we just stored 3 sequences as variables. Now we're going to compute the hamming distance between them using the function ``hamming()``. Don't forget we can look at what any function does by using ``%psource``. Let's see what the ``hamming()`` function looks like (don't worry if some parts of it aren't entirely clear just yet):
+
+```python
 >>> %psource hamming
 ```
+
+Let's find out if our query ("q1") is closer to reference 1 or reference 2 based on the hamming distance:
 
 ```python
 >>> print(hamming(r1, q1))
@@ -37,6 +46,10 @@ Here we've assumed that only *substitution events* have occurred, meaning one DN
 
 ```python
 >>> q2 = DNA("TCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCAT")
+...
+>>> print(q1)
+>>> print(q2)
+...
 >>> print(hamming(r1, q2))
 ```
 
@@ -46,10 +59,14 @@ What we'd really want to do is have a way to indicate that a deletion seems to h
 
 ```python
 >>> q3 = DNA("-TCCAGGTAAACGGTGACCAGGTACCAGTTGCGTTTGTTGTAGGAGACACGGGGACCCA")
+...
+>>> print(q1)
+>>> print(q3)
+...
 >>> print(hamming(r1, q3))
 ```
 
-What we've done here is create a pairwise alignment of ``r1`` and ``q3``. In other words, we've **aligned** positions to maximize the similarity of the two sequences, using the ``-`` to fill in spaces where one character is missing with respect to the other sequence. We refer to ``-`` characters in aligned sequences as **gap characters**, or gaps.
+What we've done here is create a pairwise alignment of ``r1`` and ``q3``. In other words, we've **aligned** positions to maximize the similarity of the two sequences, using the ``-`` to fill in spaces where one character is missing with respect to that location in the other sequence. We refer to ``-`` characters in aligned sequences as **gap characters**, or gaps.
 
 The *alignment* of these two sequences is clear if we print them  out, one on top of the other:
 
@@ -62,7 +79,7 @@ Scanning through these two sequences, we can see that they are largely identical
 
 ## What is a sequence alignment? <link src='e63a4f'/>
 
-Let's take a minute to think about sequence evolution and what a biological sequence alignment actually is. Over the course of biological evolution, a DNA sequence changes, most frequently due to random errors in replication (or the copying of a DNA sequence). These replications errors are referred to as **mutations**. Some types of mutation events that can occur are:
+Let's take a minute to think about sequence evolution and what a biological sequence alignment actually is. Over the course of biological evolution, a DNA sequence changes, most frequently due to random errors in replication (or the copying of a DNA sequence). These replication errors are referred to as **mutations**. Some types of mutation events that can occur are:
 
 * **substitutions**, where one base (or amino acid, in protein sequences) is replaced with another;
 * **insertions**, where one or more contiguous bases are inserted into a sequence;
@@ -80,13 +97,13 @@ Figure 1 illustrates how one ancestral DNA sequence (Figure 1a), over time, migh
 
[…]
