### 2018-09-01 69f017d
Update pairwise-alignment.md
typo
