### 2018-09-01 5324762
very minor edits to database-searching.md
