### 2018-09-03 804344b
ENH: add ISSUE_TEMPLATE.md fixes #269 (#270)
