+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,24 @@
+### Describe the environment
+
+What version of IAB are you reading (e.g. v0.0.1)?
+
+```python
+from iab import __version__ as iab_version
+print(iab_version)
+```
+  
+### Describe the problem
+
+#### Steps to reproduce
+
+1. ...
+2. ...
+3. ...
+  
+#### Observed Results
+
+  * What happened?  This could be a description, log output, etc.
+  
+#### Expected Results
+
+  * What did you expect to happen?
