### 2018-09-03 bebb429
MAINT: fix typos (#274)
