### 2018-09-04 a6b676c
minor suggestions/edits to multiple-sequence-alignment.md in fundamentals
