### 2018-09-21 4491be6
Update multiple-sequence-alignment.md
incorporating some more edits/wording changes, including issue [#296](https://github.com/applied-bioinformatics/An-Introduction-To-Applied-Bioinformatics/issues/296)
