### 2018-09-29 #27 @tracykteal
Thank you so much for your submission and I'm sorry for the delay.

I've read through the Getting Started section. I've also gone through installation for the interactive version, and used the materials through the Binder link. Documentation about how to use the resources and even report issues is well documented and everything works. I like the link to the Issue tracker for reporting issues, already providing a mechanism for feedback and discussion that isn't 'send email'. These resources meet the model of being both open educational materials and taking an open source approach.

@labarba this should be officially accepted and published. Thanks!

@gregcaporaso can you create an archive (on Zenodo, figshare, or other) and post the archive DOI here. Then we link to that DOI for the publication.

Thanks so much for this paper!
