### 2018-09-29 #27 @labarba
@gregcaporaso : Can you add DOIs to the journal entries in the References list?
