### 2018-09-29 #27 @labarba
Also, the book by Dunn has a website: http://practicalcomputing.org
