### 2018-10-01 3c5ebed
adds missing info to references
