+++ paper.bib
@@ -3,7 +3,7 @@
 
 @BOOK{Dunn2010-ik,
   title     = "Practical Computing for Biologists",
-  author    = "Dunn, Casey",
+  author    = "Dunn, Casey and Haddock, Steven HD",
   abstract  = "Increasingly, scientists find themselves facing exponentially
                larger data sets and analyses without suitable tools to deal
                with them. Many biologists end up using spreadsheet programs for
@@ -34,7 +34,8 @@
   publisher = "Sinauer Associates, Inc.",
   edition   = "First edition",
   month     =  nov,
-  year      =  2010
+  year      =  2010,
+  url       = "http://practicalcomputing.org/"
 }
 
 @ARTICLE{Wilson2016-kh,
@@ -48,7 +49,8 @@
   journal  = "F1000Res.",
   volume   =  3,
   month    =  jan,
-  year     =  2016
+  year     =  2016,
+  doi      = "https://doi.org/10.12688/F1000RESEARCH.3-62.V2"
 }
 
 @ARTICLE{Searls2014-ac,
@@ -67,7 +69,8 @@
   number   =  6,
   pages    = "e1003662",
   month    =  jun,
-  year     =  2014
+  year     =  2014,
+  doi      = "https://doi.org/10.1371/journal.pcbi.1003662"
 }
 
 @BOOK{Felsenstein2003-wm,
@@ -120,7 +123,7 @@
   publisher = "Cambridge University Press",
   edition   = "1 edition",
   month     =  may,
-  year      =  1998
+  year      =  1998,
 }
 
 @ARTICLE{Searls2012-ab,
@@ -143,7 +146,8 @@
   number   =  9,
   pages    = "e1002632",
   month    =  sep,
-  year     =  2012
+  year     =  2012,
+  doi      = "https://dx.doi.org/10.1371%2Fjournal.pcbi.1002632"
 }
 
 @MISC{rosalind-nc,
