### 2018-10-01 #27 @gregcaporaso
@labarba, I've added DOIs for all of the references that have DOIs, and I've added the website and co-author to the _Practical Computing_ reference. Thanks for spotting those!
