### 2018-10-01 #27 @labarba
There's something wrong still with the DOI links. The reference to Searls 2014, for example, links to `https://doi.org/https://doi.org/10.1371/journal.pcbi.1003662`, which does not resolve. Similar problem with Wilson 2016, plus a overfull line. Pls try to fix.
