### 2018-10-01 #27 @gregcaporaso
Ok, I _think_ that change should fix it. I had full URLs for these DOIs (e.g., `https://doi.org/10.12688/F1000RESEARCH.3-62.V2`) where I should have just had the DOI itself (e.g., `10.12688/F1000RESEARCH.3-62.V2`. Sorry about that!
