### 2018-10-01 #27 @labarba
I just downloaded the paper PDF, checked the links to the DOIs, and they're still broken.
