### 2018-10-02 7db02b9
remove domain from DOIs (#300)
