+++ paper.bib
@@ -50,7 +50,7 @@
   volume   =  3,
   month    =  jan,
   year     =  2016,
-  doi      = "https://doi.org/10.12688/F1000RESEARCH.3-62.V2"
+  doi      = "10.12688/F1000RESEARCH.3-62.V2"
 }
 
 @ARTICLE{Searls2014-ac,
@@ -70,7 +70,7 @@
   pages    = "e1003662",
   month    =  jun,
   year     =  2014,
-  doi      = "https://doi.org/10.1371/journal.pcbi.1003662"
+  doi      = "10.1371/journal.pcbi.1003662"
 }
 
 @BOOK{Felsenstein2003-wm,
@@ -147,7 +147,7 @@
   pages    = "e1002632",
   month    =  sep,
   year     =  2012,
-  doi      = "https://dx.doi.org/10.1371%2Fjournal.pcbi.1002632"
+  doi      = "10.1371/journal.pcbi.1002632"
 }
 
 @MISC{rosalind-nc,
