### 2018-10-02 8613ff2
fix rendering of references
