+++ paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-_Statement of need_: Due to the increasing rate of biological data generation, bioinformatics is rapidly growing as a field and is now an essential part of scientific advances in human health and environmental sciences. Online and publicly accessible resources for learning bioinformatics exist [@Searls2014-ac; @rosalind-nc], and there are excellent textbooks in the area, some focused heavily on theory [@Felsenstein2003-wm; @Durbin1998-ru], and others geared toward learning specific skills such as Python programming or the Unix shell [@Dunn2010-ik; @Wilson2016-kh]. An Introduction to Applied Bioinformatics (IAB) is a free, online bioinformatics text that bridges the gap between theory and application by teaching fundamentals of bioinformatics in the context of their implementation, using an interactive framework based on highly relevant tools including Python 3, Jupyter Notebooks, and GitHub.
+_Statement of need_: Due to the increasing rate of biological data generation, bioinformatics is rapidly growing as a field and is now an essential part of scientific advances in human health and environmental sciences. Online and publicly accessible resources for learning bioinformatics exist (e.g., [Rosalind](http://rosalind.info), [@Searls2012-ab; @Searls2014-ac]), and there are excellent textbooks and courses in the area, some focused heavily on theory [@Felsenstein2003-wm; @Durbin1998-ru], and others geared toward learning specific skills such as Python programming or the Unix shell [@Dunn2010-ik; @Wilson2016-kh]. An Introduction to Applied Bioinformatics (IAB) is a free, online bioinformatics text that bridges the gap between theory and application by teaching fundamentals of bioinformatics in the context of their implementation, using an interactive framework based on highly relevant tools including Python 3, Jupyter Notebooks, and GitHub.
 
 IAB is geared toward students who are completely new to bioinformatics, though having completed an introductory course (or book) in both Computer Science and Biology are useful prerequisites. IAB readers begin on the [project website](http://readIAB.org). While it is possible to view the content statically from this page, we recommend that readers work interactively by installing IAB. Readers progress through chapters that introduce fundamental topics, such as sequence homology searching and multiple sequence alignment, and presents their Python 3 implementation. Because the content is presented in Jupyter Notebooks, students can edit and execute the code, for example to explore how changing k-word size or an alignment gap penalty might impact the results of a database search. The Python code that readers interact with is intended for educational purposes, where the implementation is made as simple as possible, sometimes at the cost of computational efficiency. Chapters therefore also include examples of performing the same analyses with [scikit-bio](http://scikit-bio.org), a production-quality bioinformatics Python 3 library. This enables a rapid transition from learning theory, or how an algorithm works, to applying techniques in a real-world setting.
 
