+++ paper.bib
@@ -1,3 +1,4 @@
+
 % Generated by Paperpile. Check out http://paperpile.com for more information.
 % BibTeX export options can be customized via Settings -> BibTeX.
 
@@ -46,7 +47,7 @@
               worldwide volunteer effort to improve researchers' computing
               skills. This paper explains what we have learned along the way,
               the challenges we now face, and our plans for the future.",
-  journal  = "F1000Res.",
+  journal  = "F1000 Research",
   volume   =  3,
   month    =  jan,
   year     =  2016,
@@ -64,7 +65,7 @@
               bioinformatics and computational biology. The courses are
               organized into eleven subject areas modeled on university
               departments and are accompanied by commentary and career advice.",
-  journal  = "PLoS Comput. Biol.",
+  journal  = "PLoS Computational Biology",
   volume   =  10,
   number   =  6,
   pages    = "e1003662",
@@ -95,7 +96,7 @@
                fundamental matrix algebra), molecular sequences, and
                quantitative genetics.",
   publisher = "Sinauer Associates",
-  edition   = "2 edition",
+  edition   = "Second edition",
   month     =  sep,
   year      =  2003
 }
@@ -121,7 +122,7 @@
                at the same time presents the state of the art in this new and
                important field.",
   publisher = "Cambridge University Press",
-  edition   = "1 edition",
+  edition   = "First edition",
   month     =  may,
   year      =  1998,
 }
@@ -141,17 +142,11 @@
               in the form of a virtual course catalog, together with editorial
               commentary, and an assessment of strengths, weaknesses, and
               likely future directions for open online learning in this field.",
-  journal  = "PLoS Comput. Biol.",
+  journal  = "PLoS Computational Biology",
   volume   =  8,
   number   =  9,
   pages    = "e1002632",
   month    =  sep,
   year     =  2012,
   doi      = "10.1371/journal.pcbi.1002632"
-}
-
-@MISC{rosalind-nc,
-  title        = "{ROSALIND}",
-  howpublished = "\url{http://rosalind.info}",
-  note         = "Accessed: 07-17-2018"
 }
