### 2018-10-02 #27 @labarba
Thanks. The author list on the Zenodo entry does not match the author list on the paper. Zenodo auto-generates the author list from the committers on the repo, so you may need to manually edit the entry.
