### 2018-10-02 #27 @gregcaporaso
> The author list on the Zenodo entry does not match the author list on the paper.

That's fixed now, thanks!
