### 2018-10-02 #27 @labarba
@AstrobioMike — Thank you for your review!
@tracykteal — Thanks for handling this submission as editor!
🙏

@gregcaporaso — Your paper is now published in JOSS! 🎉
