### 2018-10-02 #27 @gregcaporaso
Thank you @AstrobioMike, @tracykteal, and @labarba - it was great working with all of you on this!
